tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.invalid-calls.save-endLayer.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.invalid-calls.save-endLayer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.invalid-calls.save-endLayer</h1>
<p class="desc">Raises exception on save() + endLayer().</p>


<script>
var t = async_test("Raises exception on save() + endLayer().");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_dom("INVALID_STATE_ERR", function() {
    ctx.save();
    ctx.endLayer();
  });
  t.done();

});
</script>
